+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 0.2.0 (2022-03-27)
 ------------------
 
+* BREAKING: renamed `kwh_tax` to `kwh_tariffs` - as that is what it is :-)
 * Fixed an error on forecast.price() when the API doesn't return all keys
 * Added initial work on consumption stats